😪"Phew! (...Si-er-ra)" Just completed a big overall in the data model, getting rid of the silly StepType enum and replacing it with regular Strings. This has the nice effect of letting me ditch a bunch of switch statements that were creeping up everywhere.
Still to-do:
- Make client-side recipe storage structured the same as cloud-side (i.e. Steps broken into two arrays).
- Break out function that turns Recipes into that specific storage structure, then returns it for saving either in the Cloud or App Group.
- Add "author" property to Recipes.

I want as few switch statements in View/Interface Controllers as possible — because switch statements have to be hardcoded. It'd be cool if adding types of steps was as easy as adding a new image to the app's resources.

Custom UI things like timers and unit conversion mean I can't quite get there... So I'm considering multiple ICs for the Step views. Is that best? Who knows!