Work continues on fetchDaily(), and I've adjusted fetchRecords() to return any error it gets but also handle some errors itself.
FETCHRECORDS:
I'm not sure what the proper way to divvy up error handling between fetchRecords and the various functions that call it is — but my hunch is: have fetchRecords handle errors that simply require retrying the fetch, and have the various other functions handle errors that actually mean something to them (such as a certain record not actually existing).

FETCHDAILY:
Seems to be properly querying within *some* time range, although I need to check and make sure this is actually the time range I want. Also, it didn't throw the exception I expected...