Tracked down some bugginess amongst the various Step controllers — turns out the watch crashes, hard, when you ask it to find a frame of an animation that doesn't exist!
I've been trying to clean up the visuals, to see fewer loading rings when opening a recipe. For some reason the first frame of the Mushroom animation insists on animating even when I'm trying to use it as a still — solution: make another copy of it, outside the animation folder, labeled "still." Taylor Swift is cooperating for now, but I may want to do the same to her just to be safe.