RecipePicker is now pulling from two different prototype cells, one for the daily recipe and one for saved recipes.
Animations on cells across the app has been moved to a function called animateIn, which has also been added to UITableViewCell proper as an extension. This makes for much cleaner code within the tableview controllers. Worth noting: the extended animateIn takes care of the fade-in effect.

Visual changes to both RecipePicker and RecipeView: I removed the status bar at the top of both.